+++ bin/miden-cli/src/commands/exec.rs
@@ -79,7 +79,7 @@ impl ExecCmd {
         let mut advice_inputs = AdviceInputs::default();
         advice_inputs.extend_map(inputs);
 
-        let tx_script = client.compile_tx_script(&program)?;
+        let tx_script = client.script_builder().compile_tx_script(&program)?;
 
         let result = client
             .execute_program(account_id, tx_script, advice_inputs, BTreeSet::new())