+++ bin/miden-cli/src/errors.rs
@@ -1,6 +1,7 @@
 use std::error::Error;
 
 use miden_client::{ClientError, keystore::KeyStoreError};
+use miden_lib::utils::ScriptBuilderError;
 use miden_objects::{AccountError, AccountIdError, AssetError, NetworkIdError};
 use miette::Diagnostic;
 use thiserror::Error;
@@ -63,6 +64,9 @@ pub enum CliError {
     #[error("parse error: {1}")]
     #[diagnostic(code(cli::parse_error), help("Check the inputs."))]
     Parse(#[source] SourceError, String),
+    #[error("script builder error")]
+    #[diagnostic(code(cli::script_builder_error))]
+    ScriptBuilder(#[from] ScriptBuilderError),
     #[error("transaction error: {1}")]
     #[diagnostic(code(cli::transaction_error))]
     Transaction(#[source] SourceError, String),